refactor: Split blockSlice.ts — extract drag handlers into blockDragSlice
Extracts the self-contained pointer-event state machine from blockSlice.ts
into two new files:
- blockDragUtils.ts: pure calcDragPosition helper (zero Zustand dependency)
- blockDragSlice.ts: StateCreator with dragState + 6 pointer-event handlers

blockSlice.ts now only contains game actions (startGame through setCanvasElement).
buildingStore.ts composes both slices. BlockSlice type alias preserved for compat.

Closes #571
