+++ gamesformykids/app/games/building/store/blockSlice.ts
@@ -1,7 +1,6 @@
 import type { StateCreator } from 'zustand';
 import type { BuildingStore } from './buildingStore';
-import type { Block, DragState, ShapeType } from '../types';
-import { CANVAS_CONFIG } from '../constants';
+import type { Block, ShapeType } from '../types';
 import {
   generateBlockId,
   getRandomPosition,
@@ -12,11 +11,12 @@ import {
   playSound,
 } from '../utils';
 
-export type BlockSlice = {
+// ── Types ─────────────────────────────────────────────────────────────────────
+
+export type BlockGameSlice = {
   isPlaying: boolean;
   blocks: Block[];
   selectedBlock: Block | null;
-  dragState: DragState;
   canvasEl: HTMLDivElement | null;
   startGame: () => void;
   createBlock: (shape: ShapeType) => void;
@@ -29,40 +29,14 @@ export type BlockSlice = {
   magicShuffle: () => void;
   saveCreation: () => void;
   setCanvasElement: (el: HTMLDivElement | null) => void;
-  handleMouseDown: (e: React.MouseEvent, block: Block) => void;
-  handleTouchStart: (e: React.TouchEvent, block: Block) => void;
-  handleMouseMove: (e: React.MouseEvent) => void;
-  handleTouchMove: (e: React.TouchEvent) => void;
-  handleMouseUp: () => void;
-  handleTouchEnd: () => void;
 };
 
-function calcDragPosition(
-  clientX: number,
-  clientY: number,
-  rect: DOMRect,
-  dragOffset: { x: number; y: number },
-  showGrid: boolean,
-) {
-  const rawX = clientX - rect.left - dragOffset.x;
-  const rawY = clientY - rect.top - dragOffset.y;
-  const clampX = Math.max(0, Math.min(rawX, rect.width - CANVAS_CONFIG.BLOCK_SIZE));
-  const clampY = Math.max(0, Math.min(rawY, rect.height - CANVAS_CONFIG.BLOCK_SIZE));
-  return {
-    x: showGrid ? Math.round(clampX / CANVAS_CONFIG.GRID_SIZE) * CANVAS_CONFIG.GRID_SIZE : clampX,
-    y: showGrid ? Math.round(clampY / CANVAS_CONFIG.GRID_SIZE) * CANVAS_CONFIG.GRID_SIZE : clampY,
-  };
-}
+// ── Slice ─────────────────────────────────────────────────────────────────────
 
-export const createBlockSlice: StateCreator<BuildingStore, [], [], BlockSlice> = (set, get) => ({
+export const createBlockSlice: StateCreator<BuildingStore, [], [], BlockGameSlice> = (set, get) => ({
   isPlaying: false,
   blocks: [],
   selectedBlock: null,
-  dragState: {
-    isDragging: false,
-    dragOffset: { x: 0, y: 0 },
-    draggedBlock: null,
-  },
   canvasEl: null,
 
   startGame: () => set({ isPlaying: true }),
@@ -152,75 +126,4 @@ export const createBlockSlice: StateCreator<BuildingStore, [], [], BlockSlice> =
   },
 
   setCanvasElement: (el) => set({ canvasEl: el }),
-
-  handleMouseDown: (e, block) => {
-    e.preventDefault();
-    const { canvasEl, blocks } = get();
-    const rect = canvasEl?.getBoundingClientRect();
-    if (!rect) return;
-    set({
-      dragState: {
-        isDragging: true,
-        dragOffset: { x: e.clientX - rect.left - block.x, y: e.clientY - rect.top - block.y },
-        draggedBlock: block,
-      },
-      blocks: [...blocks.filter((b) => b.id !== block.id), block],
-    });
-  },
-
-  handleTouchStart: (e, block) => {
-    e.preventDefault();
-    const { canvasEl, blocks } = get();
-    const rect = canvasEl?.getBoundingClientRect();
-    if (!rect) return;
-    const touch = e.touches[0];
-    set({
-      dragState: {
-        isDragging: true,
-        dragOffset: {
-          x: touch.clientX - rect.left - block.x,
-          y: touch.clientY - rect.top - block.y,
-        },
-        draggedBlock: block,
-      },
-      blocks: [...blocks.filter((b) => b.id !== block.id), block],
-    });
-  },
-
-  handleMouseMove: (e) => {
-    const { dragState, canvasEl, showGrid } = get();
-    if (!dragState.isDragging || !dragState.draggedBlock) return;
-    const rect = canvasEl?.getBoundingClientRect();
-    if (!rect) return;
-    const pos = calcDragPosition(e.clientX, e.clientY, rect, dragState.dragOffset, showGrid);
-    set((s) => ({
-      blocks: s.blocks.map((b) =>
-        b.id === dragState.draggedBlock!.id ? { ...b, ...pos } : b,
-      ),
-    }));
-  },
-
-  handleTouchMove: (e) => {
-    const { dragState, canvasEl, showGrid } = get();
-    if (!dragState.isDragging || !dragState.draggedBlock) return;
-    const rect = canvasEl?.getBoundingClientRect();
-    if (!rect) return;
-    const touch = e.touches[0];
-    const pos = calcDragPosition(touch.clientX, touch.clientY, rect, dragState.dragOffset, showGrid);
-    set((s) => ({
-      blocks: s.blocks.map((b) =>
-        b.id === dragState.draggedBlock!.id ? { ...b, ...pos } : b,
-      ),
-    }));
-  },
-
-  handleMouseUp: () => {
-    const { dragState, blocks } = get();
-    if (dragState.isDragging && dragState.draggedBlock) {
-      get().addToHistory(blocks);
-    }
-    set({ dragState: { isDragging: false, dragOffset: { x: 0, y: 0 }, draggedBlock: null } });
-  },
-
-  handleTouchEnd: () => get().handleMouseUp(),
 });
