+++ gamesformykids/app/games/building/store/buildingStore.ts
@@ -1,6 +1,7 @@
 import { COLORS, SHAPES } from '../constants';
 import { makeStore } from '@/lib/stores/createStore';
-import { createBlockSlice, type BlockSlice } from './blockSlice';
+import { createBlockSlice, type BlockGameSlice } from './blockSlice';
+import { createBlockDragSlice, type BlockDragSlice } from './blockDragSlice';
 import { createHistorySlice, type HistorySlice } from './historySlice';
 import { createParticleSlice, type ParticleSlice } from './particleSlice';
 import { createAchievementSlice, type AchievementSlice } from './achievementSlice';
@@ -8,15 +9,16 @@ import { createSettingsSlice, type SettingsSlice } from './settingsSlice';
 
 // ─── Types ────────────────────────────────────────────────────────────────────
 
-export type BuildingStore = BlockSlice & HistorySlice & ParticleSlice & AchievementSlice & SettingsSlice;
+export type BuildingStore = BlockGameSlice & BlockDragSlice & HistorySlice & ParticleSlice & AchievementSlice & SettingsSlice;
 
-// Re-export for consumers that import BuildingState / BuildingActions by name
-export type { BlockSlice as BuildingState, BlockSlice as BuildingActions };
+// Re-export aliases for backward-compat consumers
+export type { BlockGameSlice as BuildingState, BlockGameSlice as BuildingActions, BlockGameSlice as BlockSlice };
 
 // ─── Store ────────────────────────────────────────────────────────────────────
 
 export const useBuildingStore = makeStore<BuildingStore>('BuildingStore', (...a) => ({
   ...createBlockSlice(...a),
+  ...createBlockDragSlice(...a),
   ...createHistorySlice(...a),
   ...createParticleSlice(...a),
   ...createAchievementSlice(...a),
